WIP; avoid to_dense() during sum_product
Now SumProduct.apply returns a tuple with 1+n components, where n is the
length of out_labels.  The first component is in turn a tuple with n
components.  The remaining components are plain torch.Tensors.
Fortunately torch.autograd.Function seems allowed to return a tuple with
a non-Tensor component.

Much refactoring remains to be done.