Turn ProductAxis into an associative smart constructor productAxis
We want unifying `(A(2)*B(17)*C(17))` and `(D(2)*(E(17)*F(17)))`
to succeed despite the difference in parentheses.  This commit
introduces a smart constructor `productAxis`, which should be used
instead of the constructor `ProductAxis`, and which enforces these
invariants:
- The number of factors is never 1
- Each factor is never itself a `ProductAxis`

Also, `PatternedTensor.__post_init__` enforces the invariant that the
physical tensor is always `squeeze`d (in other words, there is never
`PhysicalAxis(1)`).

The functions `unify` and `antiunify` are changed to handle inputs such
as `A(2)*B(15)*C(7)` and `D(6)*E(5)*(4+()+2)`.